rename coordinates to stops
in order to add more rules
