+++ src/validate.rs
@@ -64,7 +64,7 @@ pub fn validate_and_metadata(rgtfs: gtfs_structures::RawGtfs, max_issues: usize)
                     .chain(validators::duration_distance::validate(&gtfs))
                     .chain(validators::check_name::validate(&gtfs))
                     .chain(validators::check_id::validate(&gtfs))
-                    .chain(validators::coordinates::validate(&gtfs))
+                    .chain(validators::stops::validate(&gtfs))
                     .chain(validators::route_type::validate(&gtfs))
                     .chain(validators::shapes::validate(&gtfs))
                     .chain(validators::agency::validate(&gtfs))
